requirements.txt
atlassian-python-api
beautifulsoup4
pyyaml
selenium